requirements.txt
flask
pillow
numpy
inference-sdk
pandas
scikit-learn
joblib